requirements.txt
allure-pytest==2.8.6 --trusted-host pypi.org --trusted-host files.pythonhosted.org
allure-python-commons==2.8.6 --trusted-host pypi.org --trusted-host files.pythonhosted.org
attrs==23.2.0 --trusted-host pypi.org --trusted-host files.pythonhosted.org
ceritifi==2024.6.2 --trusted-host pypi.org --trusted-host files.pythonhosted.org
cffi==1.16.0 --trusted-host pypi.org --trusted-host files.pythonhosted.org
chromedriver==2.24.1 --trusted-host pypi.org --trusted-host files.pythonhosted.org
chromedriver-py==125.0.6422.141 --trusted-host pypi.org --trusted-host files.pythonhosted.org
configparser==4.0.2 --trusted-host pypi.org --trusted-host files.pythonhosted.org
iniconfig==1.1.1 --trusted-host pypi.org --trusted-host files.pythonhosted.org
pytest==7.0.0 --trusted-host pypi.org --trusted-host files.pythonhosted.org
pytest-base-url==2.1.0 --trusted-host pypi.org --trusted-host files.pythonhosted.org
pytest-html==4.1.1 --trusted-host pypi.org --trusted-host files.pythonhosted.org
pytest-metadata==3.1.1 --trusted-host pypi.org --trusted-host files.pythonhosted.org
pytest-selenium==4.1.0 --trusted-host pypi.org --trusted-host files.pythonhosted.org
pytest-variables==3.1.0 --trusted-host pypi.org --trusted-host files.pythonhosted.org
requests==2.26.0 --trusted-host pypi.org --trusted-host files.pythonhosted.org
selenium==4.21.0 --trusted-host pypi.org --trusted-host files.pythonhosted.org
setuptools==60.2.0 --trusted-host pypi.org --trusted-host files.pythonhosted.org
urllib3==1.26.5 --trusted-host pypi.org --trusted-host files.pythonhosted.org
webdriver-manager==3.4.0 --trusted-host pypi.org --trusted-host files.pythonhosted.org
